feat: creating library structure alongside gh action
